+++ standalone-metastore/metastore-server/src/test/java/org/apache/hadoop/hive/metastore/tools/metatool/TestMetaToolTaskExecuteJDOQLQuery.java
@@ -28,8 +28,8 @@ import java.io.OutputStream;
 import java.io.PrintStream;
 import java.util.Arrays;
 
-import org.apache.hadoop.hive.metastore.ObjectStore;
 import org.apache.hadoop.hive.metastore.annotation.MetastoreUnitTest;
+import org.apache.hadoop.hive.metastore.tools.MetaToolObjectStore;
 import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
@@ -58,7 +58,7 @@ public class TestMetaToolTaskExecuteJDOQLQuery {
     String entry1 = "abc";
     String entry2 = "def";
 
-    ObjectStore mockObjectStore = Mockito.mock(ObjectStore.class);
+    MetaToolObjectStore mockObjectStore = Mockito.mock(MetaToolObjectStore.class);
     doReturn(Arrays.asList(entry1, entry2))
       .when(mockObjectStore).executeJDOQLSelect(eq(selectQuery));
 
@@ -84,7 +84,7 @@ public class TestMetaToolTaskExecuteJDOQLQuery {
   private void testUpdateQuery(long ret, String expected) throws Exception {
     String updateQuery = "update a set b = 'c'";
 
-    ObjectStore mockObjectStore = Mockito.mock(ObjectStore.class);
+    MetaToolObjectStore mockObjectStore = Mockito.mock(MetaToolObjectStore.class);
     when(mockObjectStore.executeJDOQLUpdate(updateQuery)).thenReturn(ret);
 
     MetaToolTaskExecuteJDOQLQuery t = new MetaToolTaskExecuteJDOQLQuery();
